rename file that stores the BpmnPage as not dedicated to playwright
Wasn't dedicated to puppeteer in the past as well
